% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMatchedDataset.R
\name{GenerateMatchedDataset}
\alias{GenerateMatchedDataset}
\title{Generate matched datasets of exposed and candidates matches with bootstrapping}
\usage{
GenerateMatchedDataset(
  exposed,
  candidate_matches,
  unit_of_observation,
  type_of_matching = "on variables",
  time_variable_in_exposed = NULL,
  time_variables_in_candidate_matches = NULL,
  variables_with_exact_matching = NULL,
  variables_with_range_matching = NULL,
  range_of_variables_with_range_matching = NULL,
  additional_matching_rules = NULL,
  rule_for_matching_on_dates = NULL,
  output_matching = NULL,
  seeds_for_sampling = NULL,
  sample_size_per_exposed = 1,
  methodology_for_bootstrapping = NULL,
  number_of_bootstrapping_samples = 100,
  type_of_sampling = NULL,
  exclude_sameUoO = TRUE,
  algorithm_for_matching = NULL,
  threshold = NULL,
  technical_details_of_matching = NULL,
  temporary_folder = NULL
)
}
\arguments{
\item{exposed}{The exposed data table}

\item{candidate_matches}{The candidate_matches data table}

\item{unit_of_observation}{Unit of observation}

\item{type_of_matching}{Decide the type of matching. Default to 'on variables'}

\item{time_variable_in_exposed}{Variable containing the days UoO are exposed.}

\item{time_variables_in_candidate_matches}{Vector with variables containing the start and end when a UoO are candidate matches}

\item{variables_with_exact_matching}{Vector of variable names to be matched exactly.}

\item{variables_with_range_matching}{Vector of variables to be matched in a range.}

\item{range_of_variables_with_range_matching}{List of vectors. Each vector contains the range around where to match a variable defined in variables_with_range_matching. Follow the same order as variables_with_range_matching}

\item{additional_matching_rules}{Additional rule to be used during matches}

\item{rule_for_matching_on_dates}{"exact"}

\item{output_matching}{Output folder and names. After we will append bootstrap number.}

\item{seeds_for_sampling}{Seed for sampling for reproducibility}

\item{sample_size_per_exposed}{Number of controls to be matched for each exposed. Default is 1}

\item{methodology_for_bootstrapping}{The default for this argument is NULL so no bootstrapping. Other options are "Sample exposed" and "Sample units of observations".}

\item{number_of_bootstrapping_samples}{Number of bootstrap samples to be generated if methodology_for_bootstrapping is not NULL}

\item{type_of_sampling}{With or without replacement}

\item{exclude_sameUoO}{Do not match the same UoO}

\item{algorithm_for_matching}{Algorithms to test. Possible values are "naive", ...}

\item{threshold}{Bin capacity to be used for creating bins based on the original dataset}

\item{technical_details_of_matching}{Technical details of matching}

\item{temporary_folder}{Folder where to store intermediate dataset. Default is NULL, if TRUE then the default folder is tempdir() otherwise it is possible to pass a custom path. If TRUE or custom path the function enter RAM saving mode.}
}
\description{
Generate matched datasets of exposed and candidates matches with bootstrapping
}
